requirements.txt
# Core Framework
streamlit==1.29.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2
python-dateutil==2.8.2

# Machine Learning
scikit-learn==1.4.0
lightgbm==4.3.0
joblib==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# OpenAI Integration
openai==1.10.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0

# Jupyter
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.29.0
